测试，修改若干bug

+++ src/main/java/cn/algerfan/service/impl/UserServiceImpl.java
@@ -104,6 +104,10 @@ public class UserServiceImpl extends BaseDao<User> implements UserService {
         if(userId == null || userId == 0) {
             return new Result(ResultCodeEnum.UNDELETE);
         }
+        User user = userMapper.selectByPrimaryKey(userId);
+        if(user.getRole()==200) {
+            return new Result(ResultCodeEnum.UNDELETE);
+        }
         if(userMapper.deleteByPrimaryKey(userId) == 0) {
             return new Result(ResultCodeEnum.UNDELETE);
         }
@@ -155,6 +159,7 @@ public class UserServiceImpl extends BaseDao<User> implements UserService {
         if(user==null) {
             return new Result(ResultCodeEnum.UNUPDATE);
         }
+        log.info(user);
         if(user.getRole()==100) {
             user.setRole(300);
         } else if(user.getRole()==300) {
